localwebb/backend/requirements.txt
fastapi
uvicorn
llama-index
llama-index-llms-google
llama-index-embeddings-google
llama-index-vector-stores-pinecone
pinecone-client
google-cloud-storage
python-multipart
spacy
pydantic
python-dotenv
rapidfuzz
pymupdf
pillow
tqdm
requests
